Add a hack to stop ZK server leaking file handles on shutdown.
ZooKeeperServer does not call close() on its ZKDatabase at shutdown,
meaning that the txn log and snapshot files are never closed.
This causes the cleanup of ZK server's temp dir in
TestingZooKeeperServer.close() to fail and throw an IOException.

Added a workaround using reflection to call close() on the ZKDatabase
after the ZK server has been shutdown.