CURATOR-58
EnsurePath and creatingParentsIfNeeded() were not using the ACLProvider to create parent nodes. The fix for this
is complicated by the fact that ACLProvider is in the framework module but ZKPaths is in the client module. Workaround
by introducing a InternalACLProvider and making ACLProvider a "synonym" for it. It can now be passed to ZKPaths.mkdirs().