Updated ServiceDiscoveryImpl.registerService so that PERMANENT registrations are not stored within the service.
This is to prevent two side effects:

When the ServiceDiscoveryImpl class is closed all PERMANENT registrations are deleted.
This is not desireable since callers who create PERMANENT registations should expect them to survive until explicitly deleted.

PERMANENT registrations should be stateless with respect to a particular service discovery client (assume a cluster of rest services).
It should be possible to use any of them to register/unregister.
It is not desirable to keep re-registering them whenever the client losses/reconnects to Zookeeper.