CURATOR-136 Consistent relpath handling in recipes
LeaderLatch would previously normalize a relative path in some places
but not others, causing inconsistent behavior. Make it check for an
absolute path on startup.

Additionally, scrub through the rest of the recipes and validate that
the path given is also valid before processing so that clients will see
failures early.

Modify PathUtils.validatePath to return the path given, for fluent-
style invocation, modelled on guava's Preconditions.

This closes #39