CURATOR-490
Abandon the Maven trick to copy the compatibility files from the curator-test module to this module. It's not working in IntelliJ nor Jenkins. Instead there's now a script, copyfiles.sh, that copies those files over. They shouldn't ever change again so this shouldn't be a problem. But if they do, they need to be copied and committed once more.