CURATOR-491
CURATOR-332 fixed part of this, but there's another background handler that has the same problem.

The background handler in {{refresh()}} correctly checks if the instance has been closed before processing as well as clearing watchers that might have been set in the interim. This same treatment needs to be added to the background handler in {{getDataAndStat()}}.