CURATOR-498
"Protection" has a potential bug. If the connection is lost for long enough, Curator will want to kill the session. Session deletions must be handled by the Leader ZK instance. At the same time that the session kill is being processed, Curator's protection mode handling could be calling the follower that it's connected to get the current list of children - this can be handled directly by the follower instance without needing to call the leader. So, in this scenario, the client will get a list of children that includes the ZNode that will get deleted as part of killing the session.

This bug has been in Curator since we added the protection feature to it more than 6 years ago. The only feasible fix is to set a watcher on the found ZNode so that clients can be notified if the ZNode is deleted (or no longer exists). This requires changes to any client unfortunately. All Curator recipes will be updated by existing Curator client code will need to be updated
manually.