Adds option to disable parent creation for PersistentNode
Adds a `useParentCreation` boolean flag to control the parent creation.
If this flag is set to `false`, the PersistentNode recipe won't create the underlying znodes with the `createParentContainersIfNeeded()`, meaning that the parent znodes will have to exist beforehand in order for the PersistentNode to succeed its creation.

The `useParentCreation` flag is set to `true` by default.

Signed-off-by: Paul Boutes <paul.boutes@gmail.com>