CURATOR-590: Add option to disable parent creation for PersistentNode
Adds a `useParentCreation` boolean flag to control the parent creation.

If this flag is set to `false`, the `PersistentNode` won't create the underlying znodes with the `createParentContainersIfNeeded()`, meaning that the parent znodes will have to exist beforehand in order for the `PersistentNode` to succeed its creation.

The `useParentCreation` flag is set to `true` by default.

https://issues.apache.org/jira/browse/CURATOR-590

Author: Paul Boutes <paul.boutes@elastic.co>
Author: Paul Boutes <paul.boutes@gmail.com>

Reviewers: Enrico Olivelli <eolivelli@apache.org>, Cameron McKenzie <mckenzie.cam@gmail.com>, Zili Chen <wander4096@gmail.com>, Jordan Zimmerman <jordan@jordanzimmerman.com>

Closes #380 from pboutes/CURATOR-590