Modified ZPathImpl's "isParameter()" method to properly check for parameters strings (values inside of a pair parameter delimiters).  This corrects issues with checking for path resolution status in certain cases.
Modified ZPathImpl's "resolved()" method to allow for substituted parameters to have leading path separators ("/"). This is useful in the case where a parameter is used for the first node element in the path, eliminating the need for the user to worry about the leading "/".

Added an additional unit test case for the leading "/" changes, and fixed bugs in the basic test cases that were accepting the wrong resolved state of tested ZPaths.

All unit tests for curator pass.