CURATOR-537: Fix effective path can be used as a fencing token of LeaderLatch
This is a follow up to #324.

ourPath can be modified after it's retrived in checkLeadership and before isLeader saves it by ourPath.get() again - if the connection reset and node be recreated.

To avoid handling multiple concurrent cases, this patch simply saves the last node is leader as the localOurPath so that it's always the last valid fencing token - it can be verified as invalid later, but never false valid.

Signed-off-by: tison <wander4096@gmail.com>