Remove testIsolatedThreadGroup
These tests are for deprecated TreeCache and PathChildrenCache.

The functionality of isolated thread group can be easily verified from the code change in commit 83de54888d39173acd7ddad0435f955e4ba736b1, while the tests are flaky due to ThreadGroup::destory is flawed.

The ThreadGroup::destory method is terminally deprecated in JDK16. See also:

* https://bugs.openjdk.org/browse/JDK-8219197
* https://bugs.openjdk.org/browse/JDK-8220516

And thus we remove the tests for keep CI stable.

FWIW the downstream use cases previously meet the issue in CURATOR-640 are currently happy to use 5.3.0 with the patch and no further failure report.

Signed-off-by: tison <wander4096@gmail.com>