CURATOR-673: Complete BackgroundCallback if CuratorFramework got closed
Previously, if curator framework is closed while there are background
operations, these operations are dropped with no dedicated events.

It is hard for blocking clients to stop waiting after framework closed.
`CuratorListener` seems like a circumvent. But apparently, it is hard
and I don't think this is by design.

This pr introduces a `clockLock` to synchronize framework closing and
operation queuing. So we are sure that no operations could be queued
after framework marked as closing.