CURATOR-677: Complete BackgroundCallback if sub operation failed or cancelled
Currently, some background operations use auxiliary sub operations to
complete task in case of primary conditions are not satisfied. But most
of these sub operations count only success path, so they will hang
`BackgroundCallback` if they are failed or cancelled due to framework
closed.

This is the leftover of [CURATOR-673][](#464).

[CURATOR-673]: https://issues.apache.org/jira/browse/CURATOR-673