CURATOR-729: Fix PersistentWatcher dead loop after curator closed (#520)
The dead loop is multifold:
1. `CuratorFramework::watchers` does not check `CuratorFrameworkState`
   as `getData` or others do.
2. `PersistentWatcher` loops itself through `reset` in `BackgroundCallback`.
3. Callback in `inBackground(callback).forPath(path)` is invoked
   synchronously.

This commit enforces `CuratorFrameworkState` checking also to `watchers`,
`watches`, `sync`, `reconfig` and `getConfig`.

Additionally, this commit will issue `KeeperState.Closed` to listeners
of `PersistentWatcher` when curator get closed. This is not required to
fix CURATOR-729, but will make the closing behavior consistent with
ZooKeeper. Also, I think it is good for asynchronous `Watcher`.

Refs: CURATOR-529, CURATOR-673